Adds basic detect command interface
